+++ cognite/powerops/resync/__init__.py
@@ -1,15 +1,5 @@
 from .core import DATAMODEL_ID_TO_RESYNC_NAME, MODELS_BY_NAME, apply, destroy, init, plan, validate
 from .core.echo import Echo
-from .core.main import apply2
+from .v2.main import apply2
 
-__all__ = [
-    "apply",
-    "plan",
-    "init",
-    "destroy",
-    "validate",
-    "MODELS_BY_NAME",
-    "Echo",
-    "DATAMODEL_ID_TO_RESYNC_NAME",
-    "apply2",
-]
+__all__ = ["apply", "plan", "init", "destroy", "validate", "MODELS_BY_NAME", "Echo", "DATAMODEL_ID_TO_RESYNC_NAME"]
