+++ cognite/powerops/resync/core/main.py
@@ -31,7 +31,6 @@ from cognite.powerops.resync.config import ReSyncConfig
 from cognite.powerops.resync.diff import FieldDifference, ModelDifference, ModelDifferences
 from cognite.powerops.resync.models.base import AssetModel, CDFFile, CDFSequence, DataModel, Model, SpaceId
 from cognite.powerops.resync.models.v2.powerops_models import DataModelLoader
-from cognite.powerops.resync.v2.shop_to_assets import PowerAssetImporter
 from cognite.powerops.utils.serialization import environment_variables
 
 from .cdf import get_cognite_api
@@ -366,18 +365,6 @@ def apply(
         written_changes.append(removed + added_updated)
 
     return written_changes
-
-
-def apply2(config_dir: Path, client: PowerOpsClient | None = None) -> None:
-    client = client or PowerOpsClient.from_settings()
-
-    importer = PowerAssetImporter.from_directory(config_dir / "production")
-
-    assets = importer.to_power_assets()
-
-    client.v1.upsert(assets)
-
-    print(f"Upserted {len(assets)} assets")
 
 
 def destroy(
